Fixed MainWinodow::refresh() to distinguish between JAR and wrapper script
Modern distros wrap the Java JAR in a wrapper script. The JAR file is stored
somewhere else. So we check for the file to end on '.jar'. JAR files are
passed to java, while anything else is executed as command assuming the
e.g. /usr/bin/plantuml will call jave with the JAR file.

This assumption is true for at least Ubuntu and Manjaro.